requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=1.2.0=pypi_0
appdirs=1.4.4=pypi_0
astunparse=1.6.3=pypi_0
attrs=22.1.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
biopandas=0.4.1=pypi_0
biopython=1.79=pypi_0
bioservices=1.10.1=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py38h497a2fe_1001
bzip2=1.0.8=h7f98852_4
ca-certificates=2021.10.8=ha878542_0
cachetools=5.2.0=pypi_0
cattrs=22.1.0=pypi_0
certifi=2021.10.8=py38h578d9bd_1
cffi=1.15.0=py38h7f8727e_0
charset-normalizer=2.0.9=pyhd8ed1ab_0
click=8.1.3=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
colorlog=6.7.0=pypi_0
commonmark=0.9.1=pypi_0
cryptography=36.0.0=py38h9ce1e76_0
cudatoolkit=11.1.1=h6406543_8
cycler=0.11.0=pypi_0
decorator=4.4.2=py_0
deepdiff=5.8.1=pypi_0
easydev=0.12.0=pypi_0
exceptiongroup=1.0.0rc9=pypi_0
ffmpeg=4.3=hf484d3e_0
flatbuffers=2.0.7=pypi_0
fonttools=4.28.5=pypi_0
freetype=2.10.4=h0708190_1
gast=0.4.0=pypi_0
gettext=0.19.8.1=hf34092f_1004
gevent=21.12.0=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.11.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googledrivedownloader=0.4=pyhd3deb0d_1
graphein=1.5.2=pypi_0
grequests=0.6.0=pypi_0
grpcio=1.49.1=pypi_0
h5py=3.7.0=pypi_0
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
lame=3.100=h7f98852_1001
libblas=3.9.0=1_h86c2bf4_netlib
libcblas=3.9.0=5_h92ddd45_netlib
libclang=14.0.6=pypi_0
libedit=3.1.20191231=he28a2e2_2
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1d223b6_11
libgfortran-ng=11.2.0=h69a702a_11
libgfortran5=11.2.0=h5c6108e_11
libiconv=1.16=h516909a_0
liblapack=3.9.0=5_h92ddd45_netlib
libpng=1.6.37=h21135ba_2
libstdcxx-ng=11.2.0=he4da1e4_11
libtiff=4.1.0=h2733197_1
libuv=1.42.0=h7f98852_0
littleutils=0.2.2=pypi_0
llvm-openmp=12.0.1=h4bd325d_1
loguru=0.6.0=pypi_0
lxml=4.9.1=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py38h1e0a361_2
mkl_fft=1.2.0=py38hab2c0dc_1
mkl_random=1.2.0=py38hc5bc63f_1
multipledispatch=0.6.0=pypi_0
ncurses=6.2=h58526e2_4
nettle=3.6=he412f7d_0
networkx=2.5.1=pyhd8ed1ab_0
ninja=1.10.2=h4bd325d_0
numpy=1.23.3=pypi_0
oauthlib=3.2.1=pypi_0
ogb=1.3.2=pypi_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openssl=1.1.1l=h7f98852_0
opt-einsum=3.3.0=pypi_0
ordered-set=4.1.0=pypi_0
outdated=0.2.1=pypi_0
packaging=21.3=pypi_0
pandas=1.2.5=py38h1abd341_0
perl=5.32.1=1_h7f98852_perl5
pexpect=4.8.0=pypi_0
pillow=7.1.2=py38hb39fc2d_0
pip=21.2.4=py38h06a4308_0
plotly=5.10.0=pypi_0
protobuf=3.19.5=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.10.2=pypi_0
pyg=2.0.1=py38_torch_1.8.0_cu111
pygments=2.13.0=pypi_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.6=pyhd8ed1ab_0
pysocks=1.7.1=py38h578d9bd_4
python=3.8.0=h0371630_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-louvain=0.15=pyhd3deb0d_0
python_abi=3.8=2_cp38
pytorch=1.8.0=py3.8_cuda11.1_cudnn8.0.5_0
pytorch-cluster=1.5.9=py38_torch_1.8.0_cu111
pytorch-scatter=2.0.8=py38_torch_1.8.0_cu111
pytorch-sparse=0.6.12=py38_torch_1.8.0_cu111
pytorch-spline-conv=1.2.1=py38_torch_1.8.0_cu111
pytz=2021.3=pyhd8ed1ab_0
pyyaml=5.4.1=py38h497a2fe_0
readline=7.0=h7b6447c_5
requests=2.26.0=pyhd8ed1ab_1
requests-cache=0.9.6=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=12.5.1=pypi_0
rich-click=1.5.2=pypi_0
rsa=4.9=pypi_0
scikit-learn=0.24.2=py38hdc147b9_0
scipy=1.5.3=py38hb2138dd_0
seaborn=0.12.0=pypi_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.33.0=h62c20be_0
suds-community=1.1.2=pypi_0
tenacity=8.1.0=pypi_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.10.0=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.27.0=pypi_0
termcolor=2.0.1=pypi_0
threadpoolctl=3.0.0=pyh8a188c0_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.8.0=py38
torchvision=0.9.0=py38_cu111
tqdm=4.62.3=pyhd8ed1ab_0
typing-extensions=4.3.0=pypi_0
url-normalize=1.4.3=pypi_0
urllib3=1.26.7=pyhd8ed1ab_0
vim=8.2.3582=py38pl5321h526f742_0
werkzeug=2.2.2=pypi_0
wget=3.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
wrapt=1.14.1=pypi_0
xarray=2022.6.0=pypi_0
xmltodict=0.13.0=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.6=py_0
yaml=0.2.5=h516909a_0
zipp=3.8.1=pypi_0
zlib=1.2.11=h7f8727e_4
zope-event=4.5.0=pypi_0
zope-interface=5.4.0=pypi_0
zstd=1.4.9=ha95c52a_0
